requirements.txt
django=5.0.4
mysql-connector-python=8.4.0